# docs/openapi.yaml
openapi: '3.0.2'
info:
    title: Collective IAM
    version: '1.0.0'
servers:
    - url: https://ciam.centralmind.net
paths:
    /role:
        post:
            summary: 'Create a role'
            security:
                - bearerAuth:
                      ['ciam.role.create', 'ciam.permission.grant.{FLAGS}']
            description: |
                Create a new role from a name, description, and optionally also its permission flags.<br>
                `ciam.permission.grant.FLAGS` is only checked if permissions are provided, for example if you try to
                create a role with the permissions [`chipware.ban`, `chipware.unban`] then you will need to have the
                permissions [`ciam.permission.grant.chipware.ban`, `ciam.permission.grant.chipware.unban`].
            tags:
                - role
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'The name of this role'
                                description:
                                    type: string
                                    description: 'The description of this role'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'The permissions of this role'
                            required:
                                - name
                                - description
            responses:
                '200':
                    description: 'Returns the newly created role'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Role'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed saving role to database'

        patch:
            summary: 'Update a role'
            security:
                - bearerAuth:
                      [
                          'ciam.role.update.{ROLE_ID}',
                          'ciam.permission.grant.{FLAGS}',
                          'ciam.permission.revoke.{FLAGS}',
                      ]
            tags:
                - role
            requestBody:
                description: |
                    The role to update, omitted fields wont be updated.<br>
                    `ciam.permission.grant|revoke.FLAGS` is only checked if permissions are provided, for example if you try to
                    set the permissions to [`chipware.ban`, `chipware.unban`] on a role that already has the permissions
                    [`chipware.ban`, `chipware.kick`] then you will need to have the permissions
                    [`ciam.permission.grant.chipware.unban`, `ciam.permission.revoke.chipware.kick`].
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                _id:
                                    $ref: '#/components/schemas/ObjectId'
                                    description: 'The id of the role to update'
                                name:
                                    type: string
                                    description: 'New name of this role'
                                description:
                                    type: string
                                    description: 'New description of this role'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'New permissions of this role'
                            required:
                                - _id
            responses:
                '200':
                    description: 'Returns the updated role'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Role'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'Role not found'
                '500':
                    description: 'Failed saving role to database'
    /role/{roleId}:
        get:
            summary: 'Get a role'
            description: 'Get a role by its roleId'
            security:
                - bearerAuth: ['ciam.role.get.{ROLE_ID}']
            tags:
                - role
            parameters:
                - in: path
                  name: roleId
                  required: true
                  schema:
                      $ref: '#/components/schemas/ObjectId'
                  description: 'The role ID'
            responses:
                '200':
                    description: 'Returns the role'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Role'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'Role not found'

        delete:
            security:
                - bearerAuth: ['ciam.role.delete.{ROLE_ID}']
            summary: 'Delete a role'
            description: 'Delete a role by its role ID'
            tags:
                - role
            parameters:
                - in: path
                  name: roleId
                  required: true
                  schema:
                      $ref: '#/components/schemas/ObjectId'
                  description: 'The role ID'
            responses:
                '200':
                    description: 'Returns the deleted role'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Role'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'Role not found'
    /role/list:
        get:
            summary: 'List roles'
            description: 'Get a list of all roles'
            security:
                - bearerAuth: ['ciam.role.list']
            tags:
                - role
            parameters:
                - in: query
                  name: skip
                  description: 'Number of roles to skip'
                  required: false
                  schema:
                      type: number
                      minimum: 0
                - in: query
                  name: limit
                  description: 'Maximum amount of roles to return'
                  required: false
                  schema:
                      type: number
                      minimum: 1
                      maximum: 100
            responses:
                '200':
                    description: 'Returns the roles'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Role'
                '400':
                    description: 'Invalid query parameters'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed listing roles'
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        ObjectId:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$'
...
        Flag:
            type: string
            minLength: 1
            pattern: '^(([a-z]+|\?)(\.([a-z]+|\?))*(\.\*)?)|\*$'
...
        Role:
            type: object
            description: 'A role is a collection of permission flags, and can be assigned to users'
            properties:
                _id:
                    $ref: '#/components/schemas/ObjectId'
                    description: 'The roles ID, no tthe same as a discord ID'
                name:
                    type: string
                    description: 'The name of the role'
                description:
                    type: string
                    description: 'A description of this role'
                permissions:
                    type: array
                    description: 'A list of all the permission flags this role has'
                    items:
                        $ref: '#/components/schemas/Flag'
                creator:
                    type: string
                    description: 'The user id of the user who created this role'
            required:
                - _id
                - name
                - description
                - permissions
                - creator
...